refactor: move routers from __init__ into dedicated router files
